+++ GameClient/Assets/02_Scripts/EnemyManager.cs
@@ -30,7 +30,7 @@ public class EnemyManager : MonoBehaviour, IPacketReceiver
 
     public unsafe void OnPacketReceived(Packet packet)
     {
-        byte packetId = packet.pbase.packet_id;
+        ushort packetId = packet.pbase.packet_id;
 
         switch ((E_PACKET)packetId)
         {
